Move data sources to shared
